+++ xgboost-predictor/src/main/java/biz/k11i/xgboost/gbm/GradBooster.java
@@ -48,7 +48,7 @@ public interface GradBooster extends Serializable {
      * @param ntree_limit limit the number of trees used in prediction
      * @return prediction result
      */
-    double[] predict(FVec feat, int ntree_limit);
+    float[] predict(FVec feat, int ntree_limit);
 
     /**
      * Generates a prediction for given feature vector.
@@ -60,7 +60,7 @@ public interface GradBooster extends Serializable {
      * @param ntree_limit limit the number of trees used in prediction
      * @return prediction result
      */
-    double predictSingle(FVec feat, int ntree_limit);
+    float predictSingle(FVec feat, int ntree_limit);
 
     /**
      * Predicts the leaf index of each tree. This is only valid in gbtree predictor.
